backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0